+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+* Fix autodrawing for `Konva.Transformer` when it is on a different layer
 * Fix `Konva.RegularPolygon` size calculations.
 
 ## 7.1.8