+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 7.1.9
+
 * Fix autodrawing for `Konva.Transformer` when it is on a different layer
 * Fix `Konva.RegularPolygon` size calculations.